Refactor PhotoRecordResponse and SubmissionRepository***
This commit refactors the PhotoRecordResponse class by renaming the name and description properties to name_lt and description_lt respectively. It also renames the name_en and description_en properties to name_en and description_en. Additionally, the commit adds a new property submission_id to the PhotoRecordResponse class.

In the SubmissionRepository interface, a new query has been added to retrieve PhotoRecordResponse objects with the submission UUID.